move model tests
