+++ app/test/unittest_main.py
@@ -6,7 +6,7 @@ if __name__ == '__main__':
     os.environ['TESTING'] = 'true'
 
     # import tests here to register them and they'll get run when this file is run
-    from app.test.user import UserTestCase
-    from app.test.connection import ConnectionTestCase
+    from app.test.models.user import UserTestCase
+    from app.test.models.connection import ConnectionTestCase
 
     unittest.main()
